requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
langchain-text-splitters>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pypdf